Fix Discard Generators
