+++ src/sqlancer/SQLDiscardGenerator.java
@@ -11,29 +11,7 @@ public abstract class SQLDiscardGenerator {
     protected SQLDiscardGenerator() {
     }
 
-    public static SQLQueryAdapter create(YSQLGlobalState globalState) {
-        StringBuilder sb = new StringBuilder();
-        sb.append("DISCARD ");
-        // prevent that DISCARD discards all tables (if they are TEMP tables)
-        boolean hasNonTempTables = globalState.getSchema().getDatabaseTables().stream()
-                .anyMatch(t -> t.getTableType() == TableType.STANDARD);
-        String what;
-        if (hasNonTempTables) {
-            what = Randomly.fromOptions("ALL", "PLANS", "SEQUENCES", "TEMPORARY", "TEMP");
-        } else {
-            what = Randomly.fromOptions("PLANS", "SEQUENCES");
-        }
-        sb.append(what);
-        return new SQLQueryAdapter(sb.toString(), ExpectedErrors.from("cannot run inside a transaction block")) {
-
-            @Override
-            public boolean couldAffectSchema() {
-                return canDiscardTemporaryTables(what);
-            }
-        };
-    }
-
-    private static boolean canDiscardTemporaryTables(String what) {
+    protected static boolean canDiscardTemporaryTables(String what) {
         return what.contentEquals("TEMPORARY") || what.contentEquals("TEMP") || what.contentEquals("ALL");
     }
 }
